@@ -2,13 +2,13 @@ import { Component } from '@angular/core';
 import { Platform } from '@ionic/angular';
 import { SplashScreen } from '@ionic-native/splash-screen/ngx';
 import { StatusBar } from '@ionic-native/status-bar/ngx';
-import { Storage } from '@ionic/storage';
 import { Router } from '@angular/router';
-
+import { StorageService } from './services/storageService/storage.service';
 
 @Component({
   selector: 'app-root',
-  templateUrl: 'app.component.html'
+  templateUrl: 'app.component.html',
+  styleUrls: ['app.component.scss']
 })
 export class AppComponent {
 
@@ -18,27 +18,33 @@ export class AppComponent {
     private platform: Platform,
     private splashScreen: SplashScreen,
     private statusBar: StatusBar,
-    private storage: Storage,
-    private router: Router
+    private router: Router,
+    private storageService: StorageService
   ) {
     this.initializeApp();
   }
+
+  cart = [];
+  // userSettings = [];
 
   initializeApp() {
     this.platform.ready().then(() => {
       this.statusBar.styleDefault();
       this.splashScreen.hide();
-      this.storage.get('showIntro').then((value) => {
-        if (value || value == null) {
+      this.storageService.setCart(this.cart);
+      let settings = this.storageService.getUserSettings();
+      if (settings != null) {
+        if (settings.showIntro || settings.showIntro == null) {
           this.rootPage = '/intro';
-          // console.log(this.rootPage);
           this.router.navigateByUrl(this.rootPage);
         } else {
-          this.rootPage = '/tabs';
-          // console.log(this.rootPage);
-          this.router.navigateByUrl(this.rootPage);
+          // this.rootPage = '/tabs';
+          // this.router.navigateByUrl(this.rootPage);
         }
-      });
+      } else {
+        // this.rootPage = '/tabs';
+        // this.router.navigateByUrl(this.rootPage);
+      }
     });
   }
 }
